добавленна информация о версии приложения (с костылями) исправил отображение основого цвета, когда он еще не выбрал, по умолчанию был белый цвет

+++ app/src/main/java/com/example/bankapp/domain/repository/SettingsRepository.kt
@@ -4,6 +4,6 @@ interface SettingsRepository {
     fun isDarkTheme(): Boolean
     fun setDarkTheme(enabled: Boolean)
 
-    fun getPrimaryColor(): Int
-    fun setPrimaryColor(color: Int)
+    fun getPrimaryColor(): Long
+    fun setPrimaryColor(color: Long)
 }
